refactor: 강의 스킬 가중치 타입 변경, 이에따른 Mapper 파라미터 타입 변경

+++ src/main/java/com/hlinks/domain/course/dto/CourseSkillAggregationRow.java
@@ -3,12 +3,14 @@ package com.hlinks.domain.course.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class CourseSkillAggregationRow {
 
     private Long chapterId;
     private Long skillId;
-    private Integer weight;
+    private BigDecimal weight;
     private Integer durationSeconds;
 }
